analysis_service/requirements.txt
langchain==0.3.13
langchain_core==0.3.28
langchain_experimental==0.3.4
langchain_openai==0.2.14
langgraph==0.2.60
matplotlib==3.10.0
numpy==1.26.4
pydantic==2.10.4
python-dotenv==1.0.1
fastapi
uvicorn
pandas